requirements.txt
asgiref==3.8.1
attrs==24.3.0
black==24.10.0
click==8.1.8
Django==5.1.5
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
inflection==0.5.1
isort==5.13.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
PyJWT==2.10.1
PyYAML==6.0.2
referencing==0.36.1
rpds-py==0.22.3
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
uritemplate==4.1.1
